Made graph usable for non-chronological id's
Movies have all kind of ids, and react-force-graph does *not* like that. For it to work the nodes needs to have ids chronologically and without skipping numbers. So basically now the movie id is stored in a different key, and the id used for the nodes are generated as the graph component is rendered so that it always works. Also added some comments to explain stuff.