Expanded graph functionality (see description)
You can now remove movies from the graph, either from the details modal or by right clicking on them. Also when you click on an already added movie in connected movies, it will go to that movies modal making navigation between movies faster. Also added a button in the modal to go to the movies main page (for when Artem some day adds that :pray:)